重构 event, 日志改用 stderr
