+++ event/header.go
@@ -1,0 +1,31 @@
+package event
+
+import "strconv"
+
+type Header struct {
+	Ver        string
+	Server     string
+	Serial     int
+	Pool       string
+	PoolSerial int
+	EventName  string // 事件名称
+	Len        int    // Payload长度
+}
+
+func ParseHeader(header string) (*Header, error) {
+	h := &Header{}
+	fields := parseFields(header)
+	if len(fields) == 0 {
+		return h, ErrParseHeader
+	}
+
+	h.Ver = fields["ver"]
+	h.Server = fields["server"]
+	h.Serial, _ = strconv.Atoi(fields["serial"])
+	h.Pool = fields["pool"]
+	h.PoolSerial, _ = strconv.Atoi(fields["poolserial"])
+	h.EventName = fields["eventname"]
+	h.Len, _ = strconv.Atoi(fields["len"])
+
+	return h, nil
+}
