+++ go.mod
@@ -1,9 +1,12 @@
 module github.com/ouqiang/supervisor-event-listener
 
+go 1.14
+
 require (
 	github.com/go-gomail/gomail v0.0.0-20160411212932-81ebce5c23df
-	github.com/smartystreets/goconvey v0.0.0-20190330032615-68dc04aab96a // indirect
+	github.com/pkg/errors v0.9.1
+	github.com/smartystreets/goconvey v1.6.4 // indirect
 	gopkg.in/alexcesaro/quotedprintable.v3 v3.0.0-20150716171945-2caba252f4dc // indirect
 	gopkg.in/gomail.v2 v2.0.0-20160411212932-81ebce5c23df // indirect
-	gopkg.in/ini.v1 v1.28.0
+	gopkg.in/ini.v1 v1.56.0
 )
