+++ listener/listener.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/ouqiang/supervisor-event-listener/event"
 	"github.com/ouqiang/supervisor-event-listener/listener/notify"
-	"github.com/ouqiang/supervisor-event-listener/utils/tmpfslog"
 )
 
 var (
@@ -32,58 +31,14 @@ func listen() {
 	reader := bufio.NewReader(os.Stdin)
 	for {
 		ready()
-		header, err := readHeader(reader)
-		tmpfslog.Debug("header:%+v err:%+v", header, err)
+		msg, err := event.ReadMessage(reader)
 		if err != nil {
 			failure(err)
 			continue
 		}
-		payload, err := readPayload(reader, header.Len)
-		tmpfslog.Debug("payloadL%+v err:%+v", payload, err)
-		if err != nil {
-			failure(err)
-			continue
-		}
-		// 只处理进程异常退出事件
-		notify.Push(header, payload)
 		success()
+		notify.Push(msg)
 	}
-}
-
-// 读取header
-func readHeader(reader *bufio.Reader) (*event.Header, error) {
-	// 读取Header
-	data, err := reader.ReadString('\n')
-	if err != nil {
-		return nil, err
-	}
-	// 解析Header
-	header, err := event.ParseHeader(data)
-	if err != nil {
-		return nil, err
-	}
-
-	return header, nil
-}
-
-// 读取payload
-func readPayload(reader *bufio.Reader, payloadLen int) (*event.Payload, error) {
-	// 读取payload
-	buf := make([]byte, payloadLen)
-	length, err := reader.Read(buf)
-	if err != nil {
-		return nil, err
-	}
-	if payloadLen != length {
-		return nil, ErrPayloadLength
-	}
-	// 解析payload
-	payload, err := event.ParsePayload(string(buf))
-	if err != nil {
-		return nil, err
-	}
-
-	return payload, nil
 }
 
 func ready() {
@@ -96,5 +51,5 @@ func success() {
 
 func failure(err error) {
 	fmt.Fprintln(os.Stderr, err)
-	fmt.Fprint(os.Stdout, "Result 2\nFAIL")
+	fmt.Fprint(os.Stdout, "RESULT 2\nFAIL")
 }
