+++ listener/notify/notifiable.go
@@ -1,0 +1,7 @@
+package notify
+
+import "github.com/ouqiang/supervisor-event-listener/event"
+
+type Notifiable interface {
+	Send(event.Message) error
+}
