+++ tests/supervisor-app.ini
@@ -2,6 +2,7 @@
 command=/usr/local/bin/supervisor-event-listener 
     -c /etc/supervisor-event-listener.ini
 events=
+    TICK_60,
     PROCESS_STATE_EXITED,
     PROCESS_STATE_FATAL,
     PROCESS_STATE_STOPPED,
